Fix the Server Side the Angular way example
Fixes a bug when the API returns an empty array and the table won't
display the "No data!" message. This happens because the
`this.persons` array is already initialized with an empty array
within the class, so it's not updated, and the table gets
overwritten by DataTables.

Fixes indentation on snippets to match the 2 spaces used throughout
the project.